work

+++ switch-to-wifi-client.sh
@@ -1,3 +1,9 @@
+if [ "$(curl google.com | grep 301)" = "" ];
+then
+  echo "Already a wifi client";
+  exit 0;
+fi
+
 sudo /root/.nvm/versions/node/v8.17.0/bin/node -e "require('/home/pi/orobot-firmware/scripts/switch-to-wifi-client.js').writeWPAConf()"
 sudo systemctl stop hostapd.service
 sudo systemctl stop dnsmasq.service
